Checkpoint type introductions
