+++ diff-test/core/src/sanity.ts
@@ -1,5 +1,5 @@
-import _ from 'underscore';
 import { P, C, TRUSTING_SECONDS } from './constants.js';
+import { Chain, Validator } from './common.js'
 
 /**
  * Checks that the model does not overapproximate the system
@@ -11,20 +11,14 @@ import { P, C, TRUSTING_SECONDS } from './constants.js';
  */
 class Sanity {
   // the timestamp contained in the latest trusted header
-  tLastTrustedHeader = _.object([
-    [P, 0],
-    [C, 0],
-  ]) as { provider: number; consumer: number };
+  tLastTrustedHeader = { provider: 0, consumer: 0 }
   // the timestamp of the last committed block
-  tLastCommit = _.object([
-    [P, 0],
-    [C, 0],
-  ]);
+  tLastCommit = { provider: 0, consumer: 0 }
 
   /**
    * Records the updating of a client on chain at time t
    */
-  updateClient = (chain, t) => {
+  updateClient = (chain: Chain, t: number) => {
     if (this.tLastTrustedHeader[chain] + TRUSTING_SECONDS <= t) {
       throw 'EXPIRED! - not supposed to happen. There is a mistake in how the model or generator is written.';
     }
@@ -34,14 +28,14 @@ class Sanity {
   /**
    * Records the commitment of a block on chain at time t
    */
-  commitBlock = (chain, t) => {
+  commitBlock = (chain: Chain, t: number) => {
     this.tLastCommit[chain] = t;
   };
 
   /**
    * Checks that a new validator set is sensible.
    */
-  newValSet = (vals) => {
+  newValSet = (vals: Validator[]) => {
     if (vals.length < 1) {
       throw 'EMPTY VAL SET! - not supposed to happen. There is a mistake in how the model or generator is written.';
     }
